rename ExtGrid -> ExtCheckers
The `ExtChecker` class is not pub in the Rust library, so internally to
the library there should be no confusion with `ExtCheckers`. They also
derive from different classes. In addition, grid-control is not going to
be unique to checker manipulation when I add the highlighting
functionality to showcase a win condition. This adds up to the new name
making much more sense than the old name.