add clearing game functionality
This commit adds is_full() to ExtCheckers, which returns true if all
checkers have been played. In addition, it adds drop_all_checkers(),
which removes all the checkers and adds them back to red_checkers and
yellow_checkers queues if they are red or yellow respectively. This
commit also adds the functionality of checking if the board is full and
clearing it if it is.