add ExtCheckers::find_connected_fours() + related
I added find_connected_fours(), which returns an optional HashSet of
checker coordinates that are involved in a connect-four win condition. I
also connected this functionality to ExtGame::play_column(), so the game
will be able to detect a win condition when a checker is played.
Finally, I added an is_finished property to ExtGame, so that a game over
can be "held" until the player(s) interact with the game (by clicking a
column) a second time.

The last thing to work on before initial 0.0.1 release is the highlight
functionality.