add dynamic tiles + integrate controller
I made a custom ExtTileMap class, which controls the dynamic tile map.
The initial map is controlled by the editor, while the tiles are updated
to show a check as needed. I also integrated this with the game flow, so
that solved games will show green checks over the solution.

This commit really finishes the basic graphical and logical parts of the
game. I am confident to release a 0.0.1 version after any immediate
fixes are ironed out.