remove drop_all_checkers() call from checkers drop
drop_all_checkers() was causing an aborted panic if any checkers were
still on the board, because both the Godot engine and the Rust code
would try to free the ExtChecker's. This commit fixes this by not
removing the checkers in play before freeing the ones not in play,
allowing the Godot engine to deal with the ones still in the scene tree.